Optimize the courses view page, Change course Menu
Now the course menu will first show a list of departments and then when the user selects a department, it shows courses in that department.
Also the course list view was very slow on mobile devices so I removed the su-day-shower widget (which showed what days of week that course
was held) and replaced it with a simple text.