Now place desciption files are fetched from a github repository.
I have created a github repository (ShirazUniversityCatalogueData) to hold
all data that we use for this app. Now when the app is published, students themselves can participate and complete the documents.
This commit involves a few changes in PlaceFactory which has a different format of places now and a few scripts in scripts folder which automatically download the github repository and generate the place.json file and html assets for places.