Add a framework for creating static content.
Now instead of hardcoding the static content, the content can be in the Content directory of ShirazUniversityCatalogueData repository.
The collect_content.py script walks the directory structure of this folder and adds each subfolder as a menu and each html file as a card whose title is the name of the file.
For example suppose you want to add a FAQ section to the app and also you want to caegorize the questions. You create the following directory structure inside the Content folder.

faq/
	Places/
		"what are the places?.html"
		"how to get to places?.html"
		"how to use the bus?.html"
	Classes/
		"what are the classes?.html"
		"how to sign up for classes.html"

then when you rebuild the app, inside the content submenu there is a menu named faq which itself contains two links to the places and classes menus.
they show the question as cards whose title is the name of the file and copy the content of the file in the content of the card. (you can use arbitary html inside the files)

for example:

--------------------------
what are the places?
--------------------------
this is the content of
what are the places?.html
file.
--------------------------